backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-CORS>=4.0
Flask-SocketIO>=5.3
PyJWT>=2.8
qrcode[pil]>=7.4
openpyxl>=3.1
pandas>=2.1
Pillow>=10.2
python-dotenv>=1.0
werkzeug>=3.0
cryptography>=41.0
gunicorn>=21.2
xlsxwriter>=3.1
